CODE

+++ Chapter 9/I8-PracticeQs60.c
@@ -3,7 +3,7 @@
 //Q. Make a structure to store Bank Account information of a customer of ABC Bank. Also make an alias for it.
 typedef struct bank{
     char name[100];
-    double acc;
+    long long int acc;
 }cus;
 int main(){
     printf("ABC BANK\n");
@@ -12,9 +12,9 @@ int main(){
     printf("Enter Name: ");
     gets(c1.name);
     printf("Enter Account Number: ");
-    scanf("%d", &c1.acc);
+    scanf("%lld", &c1.acc);
     printf("\n");
     printf("Name of Customer: ");
     puts(c1.name);
-    printf("Account Number: %d", c1.acc);
+    printf("Account Number: %lld", c1.acc);
 }
